Move extensions into extensions folder
