+++ src/Firehose.Web/Firehose.Web.csproj
@@ -309,13 +309,13 @@
     <Compile Include="Controllers\FeedController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Infrastructure\Logger.cs" />
-    <Compile Include="Infrastructure\GravatarHashingExtensions.cs" />
+    <Compile Include="Extensions\GravatarHashingExtensions.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
     <Compile Include="Infrastructure\IAmABlogger.cs" />
     <Compile Include="Infrastructure\IFilterMyBlogPosts.cs" />
-    <Compile Include="Infrastructure\Md5HashingExtensions.cs" />
+    <Compile Include="Extensions\Md5HashingExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Authors\JimBennett.cs" />
     <Compile Include="Authors\MarcoKuiper.cs" />
