+++ src/Firehose.Web/Views/Authors/BusinessCard.cshtml
@@ -1,4 +1,6 @@
 ﻿@using Firehose.Web.Infrastructure
+@using Firehose.Web.Extensions
+
 @model IAmACommunityMember
 
 <div class="businessCard">
